add a utility to generate a merchant payment url
